+++ GameTime/Features/GameTimer/Models/GameTimer.swift
@@ -3,7 +3,7 @@ import Foundation
 struct GameTimer: Identifiable, Hashable {
     let id: UUID
     var name: String
-    var timer: Double
+    var timer: Int
 }
 
 extension GameTimer {
